+++ packages/devtools_app_shared/lib/src/service/service_extensions.dart
@@ -6,6 +6,7 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
+import 'package:flutter/services.dart';
 
 class ServiceExtension<T> {
   ServiceExtension({
@@ -216,10 +217,8 @@ final trackRebuildWidgets = ToggleableServiceExtension<bool>(
 );
 
 final profilePlatformChannels = ToggleableServiceExtension<bool>(
-  // TODO(kenz): use ${ServicesServiceExtensions.profilePlatformChannels.name}
-  // once this enum value has existed on Flutter stable for a reasonable amount
-  // of time (6 months, or June 2024).
-  extension: '${flutterExtensionPrefix}profilePlatformChannels',
+  extension:
+      '$flutterExtensionPrefix${ServicesServiceExtensions.profilePlatformChannels.name}',
   enabledValue: true,
   disabledValue: false,
 );